# src/main/webapp/swagger.json
{"swagger":"2.0","info":{"description":"Some custom description of API.","version":"1.0","title":"DMS API","termsOfService":"https://www.power2sme.com/termsandconditions","contact":{"name":"Power2sme"},"license":{"name":"License of API","url":"https://www.power2sme.com/privacypolicy"}},"host":"localhost:8080","basePath":"/dms/api/v1","tags":[{"name":"Document Collection","description":" "}],"paths":{"/documentsinfo":{"get":{"tags":["Document Collection"],"summary":"Get All Documents for SME","description":"Returns the list of all Documents and a particular Document along with latest version, its digital info, its verification status and comments provided by the different system for SME.","operationId":"getDocumentsUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"smeId","in":"query","description":"SME Id","required":true,"type":"string","default":"ACC0008"},{"name":"docTypeId","in":"query","description":"Document Type Id","required":false,"type":"integer","format":"int32"},{"name":"docId","in":"query","description":"Document Id","required":false,"type":"integer","format":"int32"},{"name":"versionNo","in":"query","description":"Version No","required":false,"type":"integer","format":"int32"},{"name":"systemName","in":"query","description":"System Name","required":true,"type":"string","default":"fsme"}],"responses":{"404":{"description":"Not Found"},"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseDto"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}},"post":{"tags":["Document Collection"],"summary":"Update documents for list of Document types","description":"Creates List of Document Types consisting of documents which in turn consists of files, comments,digitized info.","operationId":"updateDocumentUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"documentWrapper","description":"documentWrapper","required":true,"schema":{"$ref":"#/definitions/Document Wrapper"}}],"responses":{"404":{"description":"Not Found"},"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/documenttype":{"get":{"tags":["Document Collection"],"summary":"Get All Document Type","description":"Returns the list of all Document type accesible to an application.","operationId":"getDocumentTypeUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"systemName","in":"query","description":"System Name","required":true,"type":"string","default":"fsme"}],"responses":{"404":{"description":"Not Found"},"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseDto"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/downloadfile":{"get":{"tags":["Document Collection"],"summary":"Download file from DMS System","description":"In this you can download a file by providing unique file id from the DMS System.","operationId":"downloadFileUsingGET","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"fileId","in":"query","description":"File Id","required":true,"type":"string"}],"responses":{"404":{"description":"Not Found"},"200":{"description":"OK"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/mapsmeid":{"post":{"tags":["Document Collection"],"summary":"Update unique identifier with SME ID","description":". You have to provide the unique identifier and sme id and it will upadte the unique identifier with sme id in the DMS System.","operationId":"updateSmeUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"smeIdMap","description":"smeIdMap","required":true,"schema":{"$ref":"#/definitions/SmeIdMap"}}],"responses":{"404":{"description":"Not Found"},"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/uploadfile":{"post":{"tags":["Document Collection"],"summary":"Upload file in the DMS System","description":"In this you can upload the file int the DMS System and it will return the unique file id.","operationId":"uploadDocumentUsingPOST","consumes":["multipart/form-data"],"produces":["*/*"],"parameters":[{"name":"file","in":"formData","description":"file","required":true,"type":"file"}],"responses":{"404":{"description":"Not Found"},"200":{"description":"OK","schema":{"$ref":"#/definitions/ResponseDto"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/zipfiles":{"post":{"tags":["Document Collection"],"summary":"Download Zip file from DMS System","description":"In this you can download all the files of a particular document or all document type in a document type name folder in a zip file.","operationId":"zipFileUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"zipWrapper","description":"zipWrapper","required":true,"schema":{"$ref":"#/definitions/ZipWrapper"}}],"responses":{"404":{"description":"Not Found"},"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}}},"definitions":{"DigitizedInfo":{"type":"object","properties":{"digitizedInfoId":{"type":"integer","format":"int32"},"key":{"type":"string"},"parentId":{"type":"integer","format":"int32"},"type":{"type":"string"},"value":{"type":"string"}}},"Comments":{"type":"object","properties":{"action":{"type":"string"},"comment":{"type":"string"},"commentId":{"type":"integer","format":"int32"},"createdBy":{"type":"string"},"createdTime":{"type":"string","format":"date-time"},"createdTimeStr":{"type":"string"},"documentId":{"type":"integer","format":"int32"},"systemId":{"type":"integer","format":"int32"},"systemName":{"type":"string"},"version":{"type":"integer","format":"int32"}}},"Document":{"type":"object","properties":{"comments":{"type":"array","items":{"$ref":"#/definitions/Comments"}},"createdOn":{"type":"string","format":"date-time"},"createdOnStr":{"type":"string"},"digitalInfoList":{"type":"array","items":{"$ref":"#/definitions/DigitizedInfo"}},"documentId":{"type":"integer","format":"int32"},"documentName":{"type":"string"},"documentTypeId":{"type":"integer","format":"int32"},"files":{"type":"array","items":{"type":"file"}},"id":{"type":"integer","format":"int32"},"smeId":{"type":"string"},"systemId":{"type":"integer","format":"int32"},"validTill":{"type":"string","format":"date-time"},"validTillStr":{"type":"string"},"verification":{"type":"array","items":{"$ref":"#/definitions/Verification"}},"versionList":{"type":"array","items":{"$ref":"#/definitions/Version"}},"versionNo":{"type":"integer","format":"int32"}}},"SmeIdMap":{"type":"object","properties":{"smeId":{"type":"string"},"uniqueId":{"type":"string"}}},"Verification":{"type":"object","properties":{"status":{"type":"string"},"systemId":{"type":"integer","format":"int32"},"systemName":{"type":"string"},"updatedBy":{"type":"string"}}},"File":{"type":"object","properties":{"file":{"type":"file"},"fileId":{"type":"string"},"fileName":{"type":"string"},"fileSize":{"type":"integer","format":"int64"}}},"DocumentType":{"type":"object","properties":{"applicableCompType":{"type":"string"},"categoryId":{"type":"integer","format":"int32"},"documentList":{"type":"array","items":{"$ref":"#/definitions/Document"}},"documentTypeId":{"type":"integer","format":"int32"},"documentTypeName":{"type":"string"},"systemName":{"type":"string"}}},"ZipWrapper":{"type":"object","properties":{"smeId":{"type":"string"},"smeName":{"type":"string"},"zipDocumentList":{"type":"array","items":{"$ref":"#/definitions/ZipDocumentsId"}}}},"Version":{"type":"object","properties":{"createdOn":{"type":"string","format":"date-time"},"createdOnStr":{"type":"string"},"digitalInfo":{"type":"array","items":{"$ref":"#/definitions/DigitizedInfo"}},"files":{"type":"array","items":{"type":"file"}},"verification":{"$ref":"#/definitions/Verification"},"versionNo":{"type":"integer","format":"int32"}}},"ZipDocumentsId":{"type":"object","properties":{"documentId":{"type":"integer","format":"int32"},"documentTypeId":{"type":"integer","format":"int32"},"versionNo":{"type":"integer","format":"int32"}}},"Document Wrapper":{"type":"object","properties":{"docType":{"type":"array","items":{"$ref":"#/definitions/DocumentType"}},"smeId":{"type":"string"},"systemName":{"type":"string"},"updatedBy":{"type":"string"}},"description":"Model for information of Document for SME."},"ResponseDto":{"type":"object","properties":{"data":{"type":"object"},"errorCode":{"type":"integer","format":"int32"},"message":{"type":"string"},"status":{"type":"string"},"totalRecords":{"type":"integer","format":"int32"}}}}}